// src/restaurant/interfaces/survey.ts
import { Image } from '@/shared/interfaces';

export interface Survey {
  restaurant: string;
  name: string;
  description: string;
  questions: Question[];
  isActive: boolean;
  _id: string;
  createdAt: string;
  updatedAt: string;
  __v: number;
}

export interface SurveyRestaurant {
  restaurant: {
    name: string;
    logo: Image;
    banner: Image;
  };
  name: string;
  description: string;
  questions: QuestionAnswer[];
  isActive: boolean;
  _id: string;
  createdAt: string;
  updatedAt: string;
  __v: number;
}

export type TypeAnswer = 'text' | 'number' | 'date' | 'time' | 'email' | 'textarea' | 'phone' | 'radio' | 'checkbox';

export interface Question {
  question: string;
  typeAnswer: TypeAnswer;
  typeQuestion: 'open' | 'close';
  options: Option[];
  _id: string;
}

export interface Option {
  _id: string;
  option: string;
}

export interface QuestionAnswer {
  question: string;
  typeAnswer: TypeAnswer;
  typeQuestion: 'open' | 'close';
  options: OptionAnswer[];
  answer: string;
  _id: string;
}

export interface OptionAnswer {
  _id: string;
  option: string;
  checked: boolean;
}

/*  */

// Generated by https://quicktype.io

export interface SurveyResolved {
  _id: string;
  restaurant: string;
  client: Client;
  questions: QuestionAnswer[];
  createdAt: string;
  updatedAt: string;
  __v: number;
}

export interface Client {
  _id: string;
  name: string;
  lastname: string;
  photo: Image | null;
  city: string;
  country: string;
  email: string;
}